application-code/kafka-fake-data-endpoint/requirements.txt
flask 
kafka-python 
pydantic
gunicorn
